+++ include/ramulator/base.h
@@ -8,8 +8,8 @@
 #include <functional>
 #include <iostream>
 
-#include <spdlog/spdlog.h>
-#include <yaml-cpp/yaml.h>
+#include "spdlog/spdlog.h"
+#include "yaml-cpp/yaml.h"
 
 #include "base/type.h"
 #include "base/factory.h"
@@ -18,7 +18,7 @@
 #include "base/exception.h"
 #include "base/logging.h"
 #include "base/debug.h"
-#include "ramulator/base/request.h"
+#include "request.h"
 #include "base/utils.h"
 #include "base/stats.h"
