+++ include/ramulator/frontend.h
@@ -5,8 +5,8 @@
 #include <string>
 #include <functional>
 
-#include "ramulator/base/base.h"
-#include "ramulator/memory_system/memory_system.h"
+#include "ramulator/base.h"
+#include "memory_system.h"
 
 
 namespace Ramulator {
