+++ src/addr_mapper/addr_mapper.h
@@ -4,7 +4,7 @@
 #include <spdlog/spdlog.h>
 #include <yaml-cpp/yaml.h>
 
-#include "ramulator/base/base.h"
+#include "ramulator/base.h"
 #include "dram_controller/controller.h"
 
 namespace Ramulator {
