+++ src/addr_mapper/impl/linear_mappers.cpp
@@ -1,9 +1,9 @@
 #include <vector>
 
-#include "ramulator/base/base.h"
+#include "ramulator/base.h"
 #include "dram/dram.h"
 #include "addr_mapper/addr_mapper.h"
-#include "ramulator/memory_system/memory_system.h"
+#include "ramulator/memory_system.h"
 
 namespace Ramulator {
