+++ src/base/factory.cpp
@@ -2,8 +2,8 @@
 #include <queue>
 
 #include "base/factory.h"
-#include "ramulator/frontend/frontend.h"
-#include "ramulator/memory_system/memory_system.h"
+#include "ramulator/frontend.h"
+#include "ramulator/memory_system.h"
 
 namespace Ramulator {
