+++ src/dram/dram.h
@@ -6,7 +6,7 @@
 #include <map>
 #include <functional>
 
-#include "ramulator/base/base.h"
+#include "ramulator/base.h"
 #include "dram/spec.h"
 #include "dram/node.h"
