+++ src/dram_controller/controller.h
@@ -7,7 +7,7 @@
 #include <spdlog/spdlog.h>
 #include <yaml-cpp/yaml.h>
 
-#include "ramulator/base/base.h"
+#include "ramulator/base.h"
 #include "dram/dram.h"
 #include "dram_controller/scheduler.h"
 #include "dram_controller/plugin.h"
