+++ src/dram_controller/impl/plugin/hydra.cpp
@@ -5,8 +5,8 @@
 #include <iomanip>
 #include <random>
 
-#include "ramulator/base/base.h"
-#include "ramulator/frontend/frontend.h"
+#include "ramulator/base.h"
+#include "ramulator/frontend.h"
 #include "translation/translation.h"
 #include "addr_mapper/addr_mapper.h"
 #include "dram_controller/controller.h"
