+++ src/dram_controller/impl/plugin/trace_recorder.cpp
@@ -7,7 +7,7 @@
 #include <spdlog/sinks/stdout_color_sinks.h>
 #include <spdlog/sinks/basic_file_sink.h>
 
-#include "ramulator/base/base.h"
+#include "ramulator/base.h"
 #include "dram_controller/controller.h"
 #include "dram_controller/plugin.h"
