+++ src/dram_controller/impl/plugin/twice.cpp
@@ -3,7 +3,7 @@
 #include <limits>
 #include <random>
 
-#include "ramulator/base/base.h"
+#include "ramulator/base.h"
 #include "dram_controller/controller.h"
 #include "dram_controller/plugin.h"
