+++ src/dram_controller/impl/refresh/all_bank_refresh.cpp
@@ -1,6 +1,6 @@
 #include <vector>
 
-#include "ramulator/base/base.h"
+#include "ramulator/base.h"
 #include "dram_controller/controller.h"
 #include "dram_controller/refresh.h"
