+++ src/dram_controller/impl/scheduler/generic_scheduler.cpp
@@ -1,6 +1,6 @@
 #include <vector>
 
-#include "ramulator/base/base.h"
+#include "ramulator/base.h"
 #include "dram_controller/controller.h"
 #include "dram_controller/scheduler.h"
