+++ src/dram_controller/refresh.h
@@ -4,7 +4,7 @@
 #include <vector>
 #include <string>
 
-#include "ramulator/base/base.h"
+#include "ramulator/base.h"
 
 
 namespace Ramulator {
