+++ src/dram_controller/scheduler.h
@@ -6,7 +6,7 @@
 #include <spdlog/spdlog.h>
 #include <yaml-cpp/yaml.h>
 
-#include "ramulator/base/base.h"
+#include "ramulator/base.h"
 
 namespace Ramulator {
