+++ src/memory_system/impl/dummy_memory_system.cpp
@@ -1,4 +1,4 @@
-#include "ramulator/memory_system/memory_system.h"
+#include "ramulator/memory_system.h"
 
 namespace Ramulator {
