+++ src/test/test_ifce.h
@@ -1,7 +1,7 @@
 #ifndef     RAMULATOR_TEST_TEST_IFCE_H
 #define     RAMULATOR_TEST_TEST_IFCE_H
 
-#include "ramulator/base/base.h"
+#include "ramulator/base.h"
 
 
 namespace Ramulator {
